+++ lib/src/model.g.dart
@@ -7,60 +7,50 @@ part of 'model.dart';
 // **************************************************************************
 
 PubScores _$PubScoresFromJson(Map<String, dynamic> json) => PubScores(
-      (json['packages'] as Map<String, dynamic>).map(
-        (k, e) => MapEntry(k, PubScore.fromJson(e as Map<String, dynamic>)),
-      ),
-      lastUpdate: json['lastUpdate'] == null
-          ? null
-          : Update.fromJson(json['lastUpdate'] as Map<String, dynamic>),
-    );
+  (json['packages'] as Map<String, dynamic>).map(
+    (k, e) => MapEntry(k, PubScore.fromJson(e as Map<String, dynamic>)),
+  ),
+  lastUpdate: json['lastUpdate'] == null
+      ? null
+      : Update.fromJson(json['lastUpdate'] as Map<String, dynamic>),
+);
 
 Map<String, dynamic> _$PubScoresToJson(PubScores instance) => <String, dynamic>{
-      'packages': instance.packages,
-      'lastUpdate': instance.lastUpdate,
-    };
+  'packages': instance.packages,
+  'lastUpdate': instance.lastUpdate,
+};
 
 PubScore _$PubScoreFromJson(Map<String, dynamic> json) => PubScore(
-      pub: PubInfo.fromJson(json['pub'] as Map<String, dynamic>),
-      github: json['github'] == null
-          ? null
-          : GitHubInfo.fromJson(json['github'] as Map<String, dynamic>),
-    );
+  pub: PubInfo.fromJson(json['pub'] as Map<String, dynamic>),
+  github: json['github'] == null
+      ? null
+      : GitHubInfo.fromJson(json['github'] as Map<String, dynamic>),
+);
 
-Map<String, dynamic> _$PubScoreToJson(PubScore instance) {
-  final val = <String, dynamic>{
-    'pub': instance.pub,
-  };
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('github', instance.github);
-  return val;
-}
+Map<String, dynamic> _$PubScoreToJson(PubScore instance) => <String, dynamic>{
+  'pub': instance.pub,
+  'github': ?instance.github,
+};
 
 PubInfo _$PubInfoFromJson(Map<String, dynamic> json) => PubInfo(
-      likeCount: json['likeCount'] as int,
-      grantedPoints: json['grantedPoints'] as int?,
-      popularity: json['popularity'] as int?,
-      lastUpdated: DateTime.parse(json['lastUpdated'] as String),
-    );
+  likeCount: (json['likeCount'] as num).toInt(),
+  grantedPoints: (json['grantedPoints'] as num?)?.toInt(),
+  downloadCount30Days: (json['downloadCount30Days'] as num?)?.toInt(),
+  lastUpdated: DateTime.parse(json['lastUpdated'] as String),
+);
 
 Map<String, dynamic> _$PubInfoToJson(PubInfo instance) => <String, dynamic>{
-      'likeCount': instance.likeCount,
-      'grantedPoints': instance.grantedPoints,
-      'popularity': instance.popularity,
-      'lastUpdated': instance.lastUpdated.toIso8601String(),
-    };
+  'likeCount': instance.likeCount,
+  'grantedPoints': ?instance.grantedPoints,
+  'downloadCount30Days': ?instance.downloadCount30Days,
+  'lastUpdated': instance.lastUpdated.toIso8601String(),
+};
 
 GitHubInfo _$GitHubInfoFromJson(Map<String, dynamic> json) => GitHubInfo(
-      json['slug'] as String,
-      starCount: json['starCount'] as int,
-      forkCount: json['forkCount'] as int,
-    );
+  json['slug'] as String,
+  starCount: (json['starCount'] as num).toInt(),
+  forkCount: (json['forkCount'] as num).toInt(),
+);
 
 Map<String, dynamic> _$GitHubInfoToJson(GitHubInfo instance) =>
     <String, dynamic>{
@@ -70,11 +60,11 @@ Map<String, dynamic> _$GitHubInfoToJson(GitHubInfo instance) =>
     };
 
 Update _$UpdateFromJson(Map<String, dynamic> json) => Update(
-      endIndex: json['endIndex'] as int,
-      date: DateTime.parse(json['date'] as String),
-    );
+  endIndex: (json['endIndex'] as num).toInt(),
+  date: DateTime.parse(json['date'] as String),
+);
 
 Map<String, dynamic> _$UpdateToJson(Update instance) => <String, dynamic>{
-      'endIndex': instance.endIndex,
-      'date': instance.date.toIso8601String(),
-    };
+  'endIndex': instance.endIndex,
+  'date': instance.date.toIso8601String(),
+};